extracting the code for a new task to a separate component
