+++ frontend/src/components/TaskList.jsx
@@ -1,38 +1,18 @@
 import PropTypes from 'prop-types';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faTrashAlt, faEdit } from '@fortawesome/free-solid-svg-icons';
+import Task from './Task';
 
-const TaskList = ({tasks, fetchTasks}) => { // the curly braces are used to destructure the props object
-    const handleDelete = async (taskId) => {
-        try {
-            const response = await fetch(`http://localhost:5213/api/tasks/${taskId}`, {
-                method: 'DELETE',
-                headers: { 
-                    'Content-Type': 'application/json'
-                },
-            });
-
-            if (!response.ok) {
-                throw new Error(`HTTP error: ${response.status}`);
-            }
-
-            fetchTasks();
-        } catch(error) {
-            console.error('Error deleting task:', error);
-        }
-    };
-    
+const TaskList = ({tasks}) => { // the curly braces are used to destructure the props object
     return (
         <>
             <ul>
                 {tasks.map((task) => (
                     <li key={task.id}>
-                        {/* <button>Reorder</button> */}
-                        <button type="delete" onClick={() => handleDelete(task.id)}>
-                            <FontAwesomeIcon icon={faTrashAlt} />
-                        </button>
-                        {/* <button><FontAwesomeIcon icon={faEdit} /></button> */}
-                        <strong>{task.title}</strong>: {task.description} - Priority: {task.priority}
+                        <Task 
+                            id={task.id}
+                            title={task.title} 
+                            description={task.description}
+                            priority={task.priority}
+                        />
                     </li>
                 ))}
             </ul>
